+++ CHANGELOG.rst
@@ -1,6 +1,10 @@
 Changelog
 ---------
 
+v1.11.dev (Nov 2021)
+    * Added `signature2plot`, adding copyright notice on plot.
+    * Added 'pyjams_color.pdf' as reference to available colormaps.
+
 v1.10 (Nov 2021)
     * Added tests for `color`.
     * Added 'pragma: no cover' to plot and MPI sections of codes so that they